Moving monitoring into its own package.
