+++ geomesa-accumulo/geomesa-accumulo-datastore/src/main/scala/org/locationtech/geomesa/accumulo/data/stats/usage/GeoMesaUsageStats.scala
@@ -13,7 +13,7 @@ import java.io.Closeable
 import org.apache.accumulo.core.client.Connector
 import org.apache.accumulo.core.security.Authorizations
 import org.joda.time.Interval
-import org.locationtech.geomesa.utils.geotools.UsageStat
+import org.locationtech.geomesa.utils.monitoring.UsageStat
 
 trait GeoMesaUsageStats extends Closeable {
