+++ geomesa-accumulo/geomesa-accumulo-datastore/src/main/scala/org/locationtech/geomesa/accumulo/data/stats/usage/UsageStatTransform.scala
@@ -16,7 +16,7 @@ import org.apache.accumulo.core.data.{Key, Mutation, Value}
 import org.apache.hadoop.io.Text
 import org.joda.time.format.DateTimeFormat
 import org.locationtech.geomesa.accumulo.util.{CloseableIterator, SelfClosingIterator}
-import org.locationtech.geomesa.utils.geotools.UsageStat
+import org.locationtech.geomesa.utils.monitoring.UsageStat
 
 import scala.util.Random
